% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_10min}
\alias{obs_10min}
\title{Data: 10-minutely meteorological observations for 'Hohenpeißenberg' site (Bavaria, Germany)}
\format{
A list of length 9 containing xts objects representing the following parameters:
\describe{
\item{PP_10}{xts object. Mean atmospheric pressure at station altitude \code{[hPa]}.}
\item{TT_10}{xts object. Mean air temperature at 2 m above ground \code{[°C]}.}
\item{TM5_10}{xts object. Mean air temperature at 5 cm above ground \code{[°C]}.}
\item{RF_10}{xts object. Mean relative humidity \code{[\%]}.}
\item{FF_10}{xts object. Mean wind velocity \code{[°C]}.}
\item{DD_10}{xts object. Mean wind direction \code{[°C]}.}
\item{GS_10}{xts object. Sum of global radiation \code{[J cm-2]}.}
\item{SD_10}{xts object. Sum of sunshine duration \code{[h]}.}
\item{RWS_10}{xts object. Precipitation depth \code{[mm]}.}
}
}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/10_minutes/}
}
\usage{
obs_10min
}
\description{
c.f. 'DESCRIPTION_obsgermany_climate_10min_***_en.pdf' files at source.
}
\details{
Historical observations of various parameters encompassing the period 1993-12-09 14:50 UTC to 2023-12-31 23:40 UTC \cr
Y = 47.8009° N; X = 11.0108° E; Z = 977 m a.s.l.
}
\note{
Last access: 2024-06-21

License: CC BY 4.0

Copyright: Deutscher Wetterdienst 2024
}
\keyword{datasets}
