% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_1d}
\alias{obs_1d}
\title{Data: Daily meteorological observations for 02290 'Hohenpeißenberg' site (Bavaria, Germany)}
\format{
A list of length 13 containing xts objects (of length 88,753 each) representing the following parameters:
\describe{
\item{FX}{xts object. Maximum windgust \code{[m s-1]}.}
\item{FM}{xts object. Mean wind velocity \code{[m s-1]}.}
\item{RSK}{xts object. Precipitation depth \code{[mm]}.}
\item{SDK}{xts object. Sunshine duration \code{[h]}.}
\item{SHK_TAG}{xts object. Snow depth \code{[cm]}.}
\item{NM}{xts object. Mean cloud cover \code{[1/8]}.}
\item{VPM}{xts object. Mean vapor pressure \code{[hPa]}.}
\item{PM}{xts object. Mean atmospheric pressure at station altitude \code{[hPa]}.}
\item{TMK}{xts object. Mean air temperature at 2 m above ground \code{[°C]}.}
\item{UPM}{xts object. Mean relative humidity \code{[\%]}.}
\item{TXK}{xts object. Maximum air temperature at 2 m above ground \code{[°C]}.}
\item{TNK}{xts object. Minimum air temperature at 2 m above ground \code{[mm]}.}
\item{TGK}{xts object. Minimum air temperature at 5 cm above ground \code{[°C]}.}
}
}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/historical/}
}
\usage{
obs_1d
}
\description{
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/daily/kl/DESCRIPTION_obsgermany-climate-daily-kl_en.pdf}
}
\details{
Historical observations of various parameters encompassing the period 1781-01-01 to 2023-12-31 \cr
Y = 47.8009° N; X = 11.0108° E; Z = 977 m a.s.l.
}
\note{
Last access: 2025-01-03

License: CC BY 4.0

Copyright: Deutscher Wetterdienst 2025
}
\keyword{datasets}
