% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_1min}
\alias{obs_1min}
\title{Data: 1-minutely meteorological observations for 'Hohenpeißenberg' site (Bavaria, Germany)}
\format{
A list of length 1 containing an xts object representing the following parameter:
\describe{
\item{RWH_RS_01}{xts object. Precipitation depth \code{[mm]}.}
}
}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/1_minute/}
}
\usage{
obs_1min
}
\description{
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/1_minute/precipitation/meta_data/}
}
\details{
Historical observations of precipitation encompassing the period 1996-01-01 00:00 UTC to 2023-12-31 22:59 UTC \cr
Y = 47.8009° N; X = 11.0108° E; Z = 977 m a.s.l.
}
\note{
Last access: 2024-06-22

License: CC BY 4.0

Copyright: Deutscher Wetterdienst 2024
}
\keyword{datasets}
